Changed to commands/manager, fixed zipping
